refactor(ResultView): add Parser class
