+++ src/main/java/lotto/ResultView.java
@@ -1,47 +1,20 @@
 package lotto;
 
-import org.w3c.dom.ranges.Range;
-
-import java.util.List;
 import java.util.Set;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 public class ResultView {
-    private static final String DELIMETER = ",";
-
-    private Numbers numbers;
+    private final Numbers numbers;
 
     public ResultView(String inputText) {
-        numbers = new Numbers(convertToNumbers(splitText(inputText)));
+        numbers = getParsedNumbers(inputText);
     }
 
     public Set<Integer> winnerNumber() {
         return numbers.getNumbers();
     }
 
-    private List<String> splitText(String text) {
-        return Stream.of(text.split(DELIMETER))
-                .map(String::trim)
-                .collect(Collectors.toList());
+    private Numbers getParsedNumbers(String inputText) {
+        return new Parser(inputText).getNumbers();
     }
 
-    private Set<Integer> convertToNumbers(List<String> text) {
-        return text.stream()
-                .map(this::parseToInt)
-                .filter(this::isLottoRange)
-                .collect(Collectors.toSet());
-    }
-
-    private boolean isLottoRange(int number) {
-        return 1 <= number && number <= 45;
-    }
-
-    private int parseToInt(String inputText) {
-        try {
-            return (Integer.parseInt(inputText));
-        } catch (Exception e) {
-            throw new IllegalArgumentException("숫자가 아닙니다. " + e.getMessage());
-        }
-    }
 }
